refactor: add minimal supported version and muzzle
