+++ instrumentation/spring/spring-cloud-gateway/spring-cloud-gateway-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/v2_0/GatewaySingletons.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.spring.gateway;
+package io.opentelemetry.javaagent.instrumentation.spring.v2_0;
 
 import io.opentelemetry.instrumentation.api.instrumenter.http.HttpServerRouteGetter;
 import org.springframework.web.server.ServerWebExchange;
