+++ settings.gradle.kts
@@ -525,7 +525,7 @@ include(":instrumentation:spring:spring-boot-autoconfigure")
 include(":instrumentation:spring:starters:spring-boot-starter")
 include(":instrumentation:spring:starters:jaeger-spring-boot-starter")
 include(":instrumentation:spring:starters:zipkin-spring-boot-starter")
-include(":instrumentation:spring:spring-cloud-gateway:javaagent")
+include(":instrumentation:spring:spring-cloud-gateway:spring-cloud-gateway-2.0:javaagent")
 include(":instrumentation:spymemcached-2.12:javaagent")
 include(":instrumentation:struts-2.3:javaagent")
 include(":instrumentation:tapestry-5.4:javaagent")
